Fix exponential walk in dead trees elimination
An earlier attempt to prevent fixUpTree() taking exponential time was
previously reverted (via commenting) due to a functional issue.

The problem with the original fix is the early return when
(node->getVisitCount() <= oldCompVisitCount). This condition will tend
to hold not only for nodes evaluated in earlier treetops, but also for
nodes that haven't been encountered yet by the current pass of dead
trees elimination, resulting in a failure to anchor.

Anchoring should often be unnecessary because of advance analysis by
isSafeToReplaceNode(), but in some cases the child of treetop is
exempted from that analysis, most notably when it has reference count 1.
In these cases it's necessary to anchor any previously unevaluated
children that might occur below. For example, consider the following
trees:

    treetop
      isub with rc=1
        iload x
        iconst -1
    istore x
      ...independent of previous value of x...
    ...
    ireturn
      ==>iload x

In order to eliminate the subtraction, the load must be anchored. With
the early return mentioned above, if the visit count of the load is
<=oldCompVisitCount (which is likely), then fixUpTree() will allow the
load to swing down past the store.

Dead trees elimination has a particular visit count value reserved to
indicate that a node is already evaluated: visitCount in process(). This
can be seen by the way visitCount is used in process() itself (after
setting future use counts), and in isSafeToReplaceNode(),
collectSymbolReferencesInNode(), and containsNode(). This commit uses
that value to identify nodes that have already been evaluated, which
never need to be anchored.

Furthermore, this commit addresses the strange usage pattern of
seenNodes, which allowed fixUpTree() to descend into the children of
nodes that it had already anchored.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>