Support call opcodes in Tril
Call opcodes are supported with this patch in tril by providing extra
annotation of the argument types and the function pointer to be invoked.

A sample looks like this:

    (method return="Int32" args=[Int32]
     (block
      (ireturn
       (icall address=0x1234 args=[Int32]
        (iload parm=0)
       ))))

which creates a method that calls another, and returns its value.

The above is included as the basis for the first test case for calls.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>