Correct ireturn child types in ShiftAndRotate tests
These tests fail on s390x due to the missing conversions before
ireturn.

The choice of doing a signed vs unsigned widening (before the `ireturn`)
is based off of how the oracle is set up right now. From where our IL
stands today, `ireturn` has no way of deciphering whether the left child
of the `*shr` (which conventionally would dictate the signedness of the
resulting shift) was signed or unsigned to begin with. So the “right choice”
here, is to mimic how the oracle interprets the resulting bits while
producing its result.

Something to keep mind, in this instance (and for future tests similar to this),
the goal is to validate whether the associated Opcode (`*shr` in this case)
behaves as expected. `ireturn` and the said conversion before it, are just
part of "extracting the computed result”. So the “extraction process” must
follow the same convention as the associated Test Oracle for the Test Case
to be deemed valid.

Fixes Issue: #1903

Related Issues: #1901

Signed-off-by: Rwitaban (Ray) Banerjee <rayb@ca.ibm.com>