Enable warnings as errors by default
Enable the warnings as errors flag for targets that are warning
free.  Other targets should be working on compiling clean for
warnings as errors and as they complete the work they should also
enable this compile flag.

Also enable -Wall for "gnu" compilers and /W3 for the msvc
compilers.  Similarly to warnings as errors add this new option
to targets that compile clean with it enabled.

Fix warnings in components that have a small number of warnings so
they can enable these new features.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>