CMake: Allow out-of-tree hookgen
Modify the hookgen handling to allow generating files outside of the
source tree. This is not trivial since the output files are specified
as a relative path in the .hdf file (which are relative to the .hdf).

To work arround this omr_add_hookgen, reads in the .hdf file, identifies
the output header names, rewrites them as relative path (relative to
the current binary dir), and then outputs the modified hdf into
CMAKE_CURRENT_BINARY_DIR, and runs hookgen on that. This allows us to
generate the hookgen files in the build tree, without making any changes
which would likely break the existing autconf/makefile builds.

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>