WIP: Fix C interface issues in omrgc.h
C interface as presented included a C++ artifact in a funciton parameter
list. The C++ function is now wrapped in a method that has an acceptable
C parameter list.

omrgc.h is a C header file and should not have included the forward
class declaration or included it in a function parameter list. However,
I have left it in the omrgc.h header file, enclosed with `#ifdef 
__cplusplus` for backwards compatibility.

Issue: #1957
Signed-off-by: Kim Briggs <briggs@ca.ibm.com>