Support building OMR with Runtime Sanitizers
Some compiler systems (gcc and clang) support the notion of Runtime
Sanitizers. These are builds that are instrumented to flag bugs at
runtime. Sanitizers are nice because they typically have a very low
false positive rate, and as such, when an issue is flagged, it typically
can be relied on.

This commit adds support for building OMR with sanitizers accessed via
the GCC or Clang command line flag `-fsanitize=`.

Closes Issue #1829

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>