Fix Compare and Branch helper for unsigned comparisons
Unsigned comparison nodes such as ifbucmpgt can be
evaluated incorrectly if the data type of the second
child is LoadConstant and 8 or 16 bits. We generate a
signed comparison instruction instead of an unsigned
instruction. This commit fixes that by correctly generating
an unsigned comparison instruction if the comparison node
is also unsigned.

Closes: #1943
Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>