Add runtime assumption with multiple patch sites
Runtime assumptions incur a persistent memory
and runtime overhead. It is possible for assumptions
to require several patch sites in a method for the same
event. It is also possible for several events to require
the same set of patch sites.

This adds a runtime assumption capable of representing
such a situation with reduced overhead, by creating
a shared collection of patch sites.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>