Disable vector reg GRA and reduction on 32bit x86
GRA does not properly assign xmm real registers to TR_VRF virtual
registers on 32bit x86. This can lead to code that produces an incorrect
result. GRA for vector registers is temporarily being disabled on 32bit
x86 until this issue is resolved.

The removal of GRA may cause a performance problem in autosimd reductions
so reductions have also been disabled. This was done by temp9rarily
disabling the use of getvelem in autosimd on 32bit x86 which is needed
for redution.

Issue: #2035
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>