Modify makefile rules for zOS
Currently the OMR zOS makefiles require that a user sets a few 
evnironment variables that control the placement of compile
options. This change uses the default placement of compile options
so that the environment variables are not required.

Remove the LIST option since we do not need to generate the 
assembly listing to stdout.

This change also removes the -g9 option that was added for
generating debug information since it is not valid.

[ci skip] - on zOS changes

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>