Co-operative copy of large objects in CS
For large objects, before trying to win forwarding, setup destination
header object to indicate that object is in progress. Then threads that
see the object is forwarded, but still being copied, will wait instead
of trying to create its own copy.

Further, while waiting, those threads will try to participate in
copying. Destination object header, beside the being-copied bit will
contantin progress info: amount of bytes yet to copy and count of
threads already participating. Threads atomically try to win a section
of object to copy. Once, everything is taken to copy (0 bytes yet to
copy), and all threads are done with the last section and decrement the
participation count, the winning threads will do the final steps of
fixup and replace the progress info with the valid object header.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>