Fail to initialize on bad limit or inline file
Currently if the JIT is unable to read the inline or limit file
supplied, it simply prints a warning and carries on. This is very
easy to miss with automated testing, and the whole purpose of
these files is for debugging and narrowing down a problem, so
there is no point to continue if the file isn't being picked up,
and worse it can lead one to false conclusions. Additionally,
the inlinefileOption accepts a range but does nothing with it.
This is obviously bad, but it makes no sense to have a range
for an inlinefile because there can easily be dependencies i.e.
one method needs to be inlined for another one to be. Therefore
the range code was removed. Doxygen documentation was also added.

Signed-off-by: Simon Hirst <shirst@ca.ibm.com>