Use Compiler Options to control behavior upon encountering an ILValidation error
Upon encountering an ILValidation failure, `continueAfterILValidationError` is
now used to decide whether we should abort compilation or not.

NOTE: The ILValidator would still report the encountered discrepencies.

Also with this, ILValidator now checks whether logging is enabled via
the `traceILValidator` compiler option (previously this was done
using the `traceILWalk` option).

Signed-off-by: Rwitaban (Ray) Banerjee <rayb@ca.ibm.com>