Build liveness with real uses on OSR points
OSRUses under prepareForOSR prolong the live range of autos and
pending push temps. Especially, those fake uses can cause many
autos and temps appear to be alive at the method entry and requires
zero initialization in the method prologue which is big overhead.

A way to solve this problem is to find out the real uses
of those autos and pending push temps at each OSRPoint and use them
during liveness analysis. The real uses can be found by figuring
out which symbols are needed to calculate the auto or temps at each
prepareForOSRCall and combining this with the liveness information
for the autos and temps from original bytecode.

This changeset changes liveness analysis to consider real uses at
each OSRPoint.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>