Fix code synchronization issue
In some cases, the store to establish the dynamically synthesized
function (i.e. create the return instruction) can still be within the
CPU core store queue. So, a stale instruction can be fetched and
executed from the iCache. This can cause an IOT/Abort trap. __lwsync()
and __isync() are used to synchronize the iCache. This prevents the
IOT/Abort trap. 

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>