Add minimum ALS for each instruction in Z codegen
Often times when implementing a new feature we make use of instructions
which have been introduced in a specific processor. To make use of such
instructions we have to guard the block of code which will emit the
respective instruction with an architecture guard. If such a guard is
omitted we end may end up executing at runtime an unsupported
instruction, depending on the environment in which the compiler is
executing.

To prevent this from happening in the future we add a minimum
architecture level set (ALS) for each instruction and add a soft assert
in the constructor of each instruction to validate that we will never
crash with a SIGILL due to a missed architecture guard.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>