Deprecate unused instruction properties in Z codegen
Some instruction properties in the Z codegen are not used and are
slated for deprecation. We currently do not envision any of the removed
properties to be used by any downstream projects now or in the future
thus eliminating stale properties allows us to minimize the size of the
instruction opcode tables and consolidate properties and properties2
static fields.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>