Remove references to libaux library
The initial OMR contribution included a library called "auxv" as part of
the port library (port/unix/auxv.[ch]). This library was written by IBM
and incorporated into the OMR code base (then the J9 JVM) long before
this code was separately released as an open source library called
libauxv. The code in OMR is, in fact, unrelated to that open source
library.

A mistake was made referencing libauxv in our license materials, however:
our LICENSE and longabout.html files reference libauxv as if it has a
different license than the rest of OMR and even mention the open source
project. This commit removes these references because they are wrong.

[ci skip] since this update does not affect the code

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>