64 bit builds should assume a 64 bit processor
This is a change targeting Power. The runtime check is64BitProcessor now
checks to see if it is a 64 build build. If so, it just returns true since
a 64 bit build must be running on a 64 bit processor. If not, the function
will check the CPU properties as before.

Issue: #2232
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>