Enable OMRChecker testing script to read CLANG environment variable
Currently, the test script always uses "clang++" to call Clang for
testing OMRChecker. This causes problems, when "clang++" calls an
unsupported clang version by default. There are a number of problems
that can arise from this, such as when using different Linux distros.
The increased flexibility was achieved by 1 change:
  - The constructor of class OMRChecker has been modified to pull and use
    Clang calling info from the environment, by looking for the env
    variable 'CLANG'. If no such env variable exists, then the default
    'clang++' is used.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>