Remove Death Tests in ILValidatorTest
The IL Validator traps when it detects invalid IL. This also
results in a core being produced. This can be wasteful when
several tril tests are purposefully testing the IL Validator by
feeding in incorrect IL. Given the above reasons it would be
better if the IL Validator threw an exception as opposed to trapping.
With this change, the Death Tests in the ILValidatorTest's tril tests
don't make sense anymore and need to be changed to instead detect the
correct return code instead. This commit adds that functionality.

Closes: #2168

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>