Replace trfprintf calls with traceMsg
trfprintf calls inside of doInstructionSelection do not do
a null check on the file handle before writing. This can cause
crashes when the file handle is null (which is a possibility
when we run out of file handles (ulimit -n)). These calls should
be replaced with traceMsg, which does a null check before trying
to print.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>