Disable vmThread reclamation on 32-bit Windows
In preparation for removing the vmThread reclamation logic, it must first
be disabled on the 32-bit platforms where it is enabled to aid with its
surgical removal:

* force `allowVMThreadRematerialization` to always return `false`.  While
this affects both Windows and Linux 32-bit, in practice this will have no
effect on Linux 32-bit because querying this setting is always coupled with
one of the other reclamation options below that mask its effect.  This
function always returns `false` on 64-bit platforms.
* force `supportsFS0VMThreadRematerialization` to always return `false`.
This affects Win32 only.
* remove calls to `setSupportsVMThreadGRA`, thereby causing queries to
`getSupportsVMThreadGRA` to always return `false`.  This affects Win32
only.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>