z/TPF system dumps segfaulting while writing system dumps.
This failure occurs when z/TPF writes out static storage
area for each of the shared objects in the VM.  z/TPF
failed to check a return code and assumed a pointer
would be valid when it wasn't.  Additionally the program
being processed no longer is loaded on z/TPF so that was
removed as well.

Signed-off-by: jjohnst-us <jjohnst@us.ibm.com>