Disable vector reg GRA and reduction on x86
Vector Reg GRA can cause a problem that leads to the generation of an
incorrect result. GRA for vector registers is temporarily being disabled
on x86 until this issue is resolved.

The removal of GRA may cause a performance problem in autosimd reductions
so reductions have also been disabled. This was done by temporarily
disabling the use of getvelem in autosimd on x86 which is needed
for redution.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>
Issue: #2280