Change the value of the unneededConv flag
The unneededConv flag used by TR::x2y nodes overlaps
with the simpleDivCheck flag. During node recreation, flags
are copied from the original node to the new node without checking
if the flag will be valid for the new node being created. This can
cause invalid instructions to be generated if the optimizer
transforms an ldiv node of two i2l children to i2l of idiv during
treeSimplification. To avoid this, this commit changes the
unneededConv node's value so that it doesn't overlap with the
simpleDivCheck flag.

Closes: #2297
Signed-off-by: Dhruv Chopra <dhruv.c.chopra@ibm.com>