Fix occasional hang in abortTest::ThreadAbortTest::Waiting
If the aborted thread has to block on it's monitor during a
monitor enter call the call to monitor enter will incorrectly
fail. The code moved 2 pieces of cleanup into 1 larger if
statement which means the second bit of cleanup is not properly
guarded.  To resolve the issue the second piece of cleanup code
needed to be guarded with the isAbortable check.

Fixes #2113

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>