Fix checks when printing RS Instructions
The getSourceImmediate() is being used incorrectly inside
TR_Debug::print(TR::FILE*,TR::S390RSInstruction*). The
purpose of the query here is to check if _sourceImmediate
is set. However, if it is set to 0, then the query evaluates
as false. This can cause errors and crashes due to printing.
This commit adds a new query to check if _sourceImmediate has
been set, and cleans up the print function to correctly use the
hasMask() query as well.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>