Make OMR::MethodBuilder client aware
Client awareness follows the approach used for IlBuilder.

The function OMRMethodBuilder::injectIL() is removed as superfluous.
All it does is to call the superclass's version and return its
return value, so we should just remove it.

OMR::MethodBuilder::AppendBuilder(TR::BytecodeBuilder *bb) is
removed in favour of OMR::MethodBuilder::AppendBytecodeBuilder().

OMR::MethodBuilder::DefineFunction() is changed to create an internal
copy of the array specifying the function parameter types.

OMR::MethodBuilder::Compile() is added. The new service provides a more
convenient interface for compiling a MethodBuilder instance, ensuring
that resources are cleaned up appropriately when compilation terminates.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>