Fixing CR Instruction handling on PPC
Prior, instructions that required CR registers were not correctly processed. An offset is needed to handle CR registers. This commit includes enabling handling for 1 Target, 2 Source Register Immediate instructions to be correctly formulated.

The implementation of the function needs to be moved from the corresponding .hpp file to the .cpp. This was necessary to determine that the register used to construct the instruction is a CR.

Signed-off-by: Alen Badel <Alen.Badel@ibm.com>