CMake: Fix DDR on Windows building using Clang
When 'clang-cl' is used as a C/C++ compiler on Windows to build OMR,
CMake attempts to link DDR with the 'pthread' library and the following
error appears:

lld-link: error: could not open pthread.lib: no such file or directory

This commit fixes the error.

In order to improve code readability, a refactoring of system
information detection was applied: a new CMake variable is introduced:
OMR_PLATFORM_THREAD_LIBRARY. A corresponding value is assigned to the
variable in a 'cmake/modules/platform/toolcfg/${OMR_TOOLCONFIG}' file.
This refactoring allows developers just to include the platform dependent
thread library to a 'target_link_libraries' invocation if the developed
component should be linked with the thread library.

Closes: #2377
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>