Build & link Port with compiler consumers (Autotools)
This commit is the first step towards leveraging more port
library features in the compiler. The port library, along
with its dependencies are set as prereqs to building the
Test Compiler or JitBuilder. The port library and its
dependencies are then linked in the product toolcfg. The
downstream projects consuming the compiler are still free
to use their own port library.

Issue: #2333

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>