Allocate a matrix in the rastest on the heap
A matrix in the method `static void matrixSquare(void)` is allocated on
the stack:

uintptr_t matrix[MATRIX_SIZE][MATRIX_SIZE]

Because the constant `MATRIX_SIZE` is defined as 1000, the amount of
memory needed to allocate `matrix` equals to 8000000. This value is more
than the default stack size for the major C/C++ toolchains. In order to
protect the test from the Stack Overflow error, the `matrix` should be
allocated on the heap as discussed in #2386.

The `omrport` library is used for memory management.

Closes: #2386
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>