Rollback on removing the ability to disable Port and Thread
Adding compiler_prereqs to targets was a way to ensure that
someone trying to build the Test Compiler or JITBuilder would
also trigger building of the port and thread libraries even
if they were set to be disabled. However, the issue with that
was that it is removing the ability to explicitly disable the
building of port and thread. Since port and thread libraries
are enabled by default, and testcompiler and jitbuilder are
disabled by default, it is safe to assume anyone enabling
the test compiler and test builder will have the port and
thread libraries available.

Also includes some nit fixes.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>