Rework the Example cmakelists
Make the default glue NOTFOUND, and assert that the glue is defined,
whenever it is required. Only use the example glue if OMR_EXAMPLE is on.
OMR_EXAMPLE is on by default.

Compile the example program. We are not able to run example as a test,
due to an old hang.

Enable OMR_FVTEST and OMR_EXAMPLE by default.

Signed-off-by: Robert Young <rwy0717@gmail.com>