Update omrsig_set_single_async_signal_handler
Changed the signature of omrsig_set_single_async_signal_handler.
Before, it returned the address to the old OS handler. Address to the
old OS handler can be NULL in non-error situations. So, it wasn't
possible to distinguish between error and non-error scenarios. Now, it
is changed to return 0 on success and non-zero on failure. The address
of the old OS handler is returned via an input reference variable.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>