Cache original OS handler once in oldActions
In Unix omrsignal.c, oldActions is used to cache the original OS
handler so that it can be restored during shutdown.
registerSignalHandlerWithOS registers a handler with the OS using
OMRSIG_SIGACTION. While registering a handler with the OS, it gets
access to the old OS handler. Now, registerSignalHandlerWithOS can be
called multiple times to register the master handlers. So, it should
cache the old (original) OS handler in oldActions while registering the
first handler. While registering subsequent handlers, oldActions should
not be updated.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>