Correct .gitignore entry for build/ directory
The entry right now says "build/" but this will match any directory
called "build" including anything below, say, jitbuilder/build . Given
the latter directory is revision controlled, it's not correct for it
to be ignored.

Fix is to use /build/ as the entry (leading slash makes it only apply
to the top level directory in the repo).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>